i/o-free python codegen
