--- rerun_py/rerun_sdk/rerun/.gitattributes
@@ -1,94 +1,0 @@
-# DO NOT EDIT! This file is generated by crates/re_types_builder/src/lib.rs
-
-.gitattributes linguist-generated=true
-archetypes/__init__.py linguist-generated=true
-archetypes/annotation_context.py linguist-generated=true
-archetypes/arrows3d.py linguist-generated=true
-archetypes/asset3d.py linguist-generated=true
-archetypes/bar_chart.py linguist-generated=true
-archetypes/boxes2d.py linguist-generated=true
-archetypes/boxes3d.py linguist-generated=true
-archetypes/clear.py linguist-generated=true
-archetypes/depth_image.py linguist-generated=true
-archetypes/disconnected_space.py linguist-generated=true
-archetypes/image.py linguist-generated=true
-archetypes/line_strips2d.py linguist-generated=true
-archetypes/line_strips3d.py linguist-generated=true
-archetypes/mesh3d.py linguist-generated=true
-archetypes/pinhole.py linguist-generated=true
-archetypes/points2d.py linguist-generated=true
-archetypes/points3d.py linguist-generated=true
-archetypes/segmentation_image.py linguist-generated=true
-archetypes/tensor.py linguist-generated=true
-archetypes/text_document.py linguist-generated=true
-archetypes/text_log.py linguist-generated=true
-archetypes/time_series_scalar.py linguist-generated=true
-archetypes/transform3d.py linguist-generated=true
-archetypes/view_coordinates.py linguist-generated=true
-blueprint/__init__.py linguist-generated=true
-blueprint/auto_space_views.py linguist-generated=true
-blueprint/panel_view.py linguist-generated=true
-components/__init__.py linguist-generated=true
-components/annotation_context.py linguist-generated=true
-components/blob.py linguist-generated=true
-components/class_id.py linguist-generated=true
-components/clear_is_recursive.py linguist-generated=true
-components/color.py linguist-generated=true
-components/depth_meter.py linguist-generated=true
-components/disconnected_space.py linguist-generated=true
-components/draw_order.py linguist-generated=true
-components/half_sizes2d.py linguist-generated=true
-components/half_sizes3d.py linguist-generated=true
-components/instance_key.py linguist-generated=true
-components/keypoint_id.py linguist-generated=true
-components/line_strip2d.py linguist-generated=true
-components/line_strip3d.py linguist-generated=true
-components/material.py linguist-generated=true
-components/media_type.py linguist-generated=true
-components/mesh_properties.py linguist-generated=true
-components/out_of_tree_transform3d.py linguist-generated=true
-components/pinhole_projection.py linguist-generated=true
-components/position2d.py linguist-generated=true
-components/position3d.py linguist-generated=true
-components/radius.py linguist-generated=true
-components/resolution.py linguist-generated=true
-components/rotation3d.py linguist-generated=true
-components/scalar.py linguist-generated=true
-components/scalar_scattering.py linguist-generated=true
-components/tensor_data.py linguist-generated=true
-components/text.py linguist-generated=true
-components/text_log_level.py linguist-generated=true
-components/transform3d.py linguist-generated=true
-components/vector3d.py linguist-generated=true
-components/view_coordinates.py linguist-generated=true
-datatypes/__init__.py linguist-generated=true
-datatypes/angle.py linguist-generated=true
-datatypes/annotation_info.py linguist-generated=true
-datatypes/class_description.py linguist-generated=true
-datatypes/class_description_map_elem.py linguist-generated=true
-datatypes/class_id.py linguist-generated=true
-datatypes/float32.py linguist-generated=true
-datatypes/keypoint_id.py linguist-generated=true
-datatypes/keypoint_pair.py linguist-generated=true
-datatypes/mat3x3.py linguist-generated=true
-datatypes/mat4x4.py linguist-generated=true
-datatypes/material.py linguist-generated=true
-datatypes/mesh_properties.py linguist-generated=true
-datatypes/quaternion.py linguist-generated=true
-datatypes/rgba32.py linguist-generated=true
-datatypes/rotation3d.py linguist-generated=true
-datatypes/rotation_axis_angle.py linguist-generated=true
-datatypes/scale3d.py linguist-generated=true
-datatypes/tensor_buffer.py linguist-generated=true
-datatypes/tensor_data.py linguist-generated=true
-datatypes/tensor_dimension.py linguist-generated=true
-datatypes/transform3d.py linguist-generated=true
-datatypes/translation_and_mat3x3.py linguist-generated=true
-datatypes/translation_rotation_scale3d.py linguist-generated=true
-datatypes/utf8.py linguist-generated=true
-datatypes/uvec2d.py linguist-generated=true
-datatypes/uvec3d.py linguist-generated=true
-datatypes/uvec4d.py linguist-generated=true
-datatypes/vec2d.py linguist-generated=true
-datatypes/vec3d.py linguist-generated=true
-datatypes/vec4d.py linguist-generated=true
