+++ rerun_py/rerun_sdk/rerun/archetypes/.gitattributes
@@ -1,0 +1,27 @@
+# DO NOT EDIT! This file is generated by crates/re_types_builder/src/lib.rs
+
+.gitattributes linguist-generated=true
+__init__.py linguist-generated=true
+annotation_context.py linguist-generated=true
+arrows3d.py linguist-generated=true
+asset3d.py linguist-generated=true
+bar_chart.py linguist-generated=true
+boxes2d.py linguist-generated=true
+boxes3d.py linguist-generated=true
+clear.py linguist-generated=true
+depth_image.py linguist-generated=true
+disconnected_space.py linguist-generated=true
+image.py linguist-generated=true
+line_strips2d.py linguist-generated=true
+line_strips3d.py linguist-generated=true
+mesh3d.py linguist-generated=true
+pinhole.py linguist-generated=true
+points2d.py linguist-generated=true
+points3d.py linguist-generated=true
+segmentation_image.py linguist-generated=true
+tensor.py linguist-generated=true
+text_document.py linguist-generated=true
+text_log.py linguist-generated=true
+time_series_scalar.py linguist-generated=true
+transform3d.py linguist-generated=true
+view_coordinates.py linguist-generated=true
