+++ rerun_py/rerun_sdk/rerun/blueprint/.gitattributes
@@ -1,0 +1,6 @@
+# DO NOT EDIT! This file is generated by crates/re_types_builder/src/lib.rs
+
+.gitattributes linguist-generated=true
+__init__.py linguist-generated=true
+auto_space_views.py linguist-generated=true
+panel_view.py linguist-generated=true
