+++ rerun_py/rerun_sdk/rerun/components/.gitattributes
@@ -1,0 +1,36 @@
+# DO NOT EDIT! This file is generated by crates/re_types_builder/src/lib.rs
+
+.gitattributes linguist-generated=true
+__init__.py linguist-generated=true
+annotation_context.py linguist-generated=true
+blob.py linguist-generated=true
+class_id.py linguist-generated=true
+clear_is_recursive.py linguist-generated=true
+color.py linguist-generated=true
+depth_meter.py linguist-generated=true
+disconnected_space.py linguist-generated=true
+draw_order.py linguist-generated=true
+half_sizes2d.py linguist-generated=true
+half_sizes3d.py linguist-generated=true
+instance_key.py linguist-generated=true
+keypoint_id.py linguist-generated=true
+line_strip2d.py linguist-generated=true
+line_strip3d.py linguist-generated=true
+material.py linguist-generated=true
+media_type.py linguist-generated=true
+mesh_properties.py linguist-generated=true
+out_of_tree_transform3d.py linguist-generated=true
+pinhole_projection.py linguist-generated=true
+position2d.py linguist-generated=true
+position3d.py linguist-generated=true
+radius.py linguist-generated=true
+resolution.py linguist-generated=true
+rotation3d.py linguist-generated=true
+scalar.py linguist-generated=true
+scalar_scattering.py linguist-generated=true
+tensor_data.py linguist-generated=true
+text.py linguist-generated=true
+text_log_level.py linguist-generated=true
+transform3d.py linguist-generated=true
+vector3d.py linguist-generated=true
+view_coordinates.py linguist-generated=true
