+++ rerun_py/rerun_sdk/rerun/datatypes/.gitattributes
@@ -1,0 +1,34 @@
+# DO NOT EDIT! This file is generated by crates/re_types_builder/src/lib.rs
+
+.gitattributes linguist-generated=true
+__init__.py linguist-generated=true
+angle.py linguist-generated=true
+annotation_info.py linguist-generated=true
+class_description.py linguist-generated=true
+class_description_map_elem.py linguist-generated=true
+class_id.py linguist-generated=true
+float32.py linguist-generated=true
+keypoint_id.py linguist-generated=true
+keypoint_pair.py linguist-generated=true
+mat3x3.py linguist-generated=true
+mat4x4.py linguist-generated=true
+material.py linguist-generated=true
+mesh_properties.py linguist-generated=true
+quaternion.py linguist-generated=true
+rgba32.py linguist-generated=true
+rotation3d.py linguist-generated=true
+rotation_axis_angle.py linguist-generated=true
+scale3d.py linguist-generated=true
+tensor_buffer.py linguist-generated=true
+tensor_data.py linguist-generated=true
+tensor_dimension.py linguist-generated=true
+transform3d.py linguist-generated=true
+translation_and_mat3x3.py linguist-generated=true
+translation_rotation_scale3d.py linguist-generated=true
+utf8.py linguist-generated=true
+uvec2d.py linguist-generated=true
+uvec3d.py linguist-generated=true
+uvec4d.py linguist-generated=true
+vec2d.py linguist-generated=true
+vec3d.py linguist-generated=true
+vec4d.py linguist-generated=true
