--- rerun_py/tests/test_types/.gitattributes
@@ -1,42 +1,0 @@
-# DO NOT EDIT! This file is generated by crates/re_types_builder/src/lib.rs
-
-.gitattributes linguist-generated=true
-archetypes/__init__.py linguist-generated=true
-archetypes/affix_fuzzer1.py linguist-generated=true
-archetypes/affix_fuzzer2.py linguist-generated=true
-archetypes/affix_fuzzer3.py linguist-generated=true
-archetypes/affix_fuzzer4.py linguist-generated=true
-blueprint/__init__.py linguist-generated=true
-components/__init__.py linguist-generated=true
-components/affix_fuzzer1.py linguist-generated=true
-components/affix_fuzzer10.py linguist-generated=true
-components/affix_fuzzer11.py linguist-generated=true
-components/affix_fuzzer12.py linguist-generated=true
-components/affix_fuzzer13.py linguist-generated=true
-components/affix_fuzzer14.py linguist-generated=true
-components/affix_fuzzer15.py linguist-generated=true
-components/affix_fuzzer16.py linguist-generated=true
-components/affix_fuzzer17.py linguist-generated=true
-components/affix_fuzzer18.py linguist-generated=true
-components/affix_fuzzer19.py linguist-generated=true
-components/affix_fuzzer2.py linguist-generated=true
-components/affix_fuzzer20.py linguist-generated=true
-components/affix_fuzzer21.py linguist-generated=true
-components/affix_fuzzer3.py linguist-generated=true
-components/affix_fuzzer4.py linguist-generated=true
-components/affix_fuzzer5.py linguist-generated=true
-components/affix_fuzzer6.py linguist-generated=true
-components/affix_fuzzer7.py linguist-generated=true
-components/affix_fuzzer8.py linguist-generated=true
-components/affix_fuzzer9.py linguist-generated=true
-datatypes/__init__.py linguist-generated=true
-datatypes/affix_fuzzer1.py linguist-generated=true
-datatypes/affix_fuzzer2.py linguist-generated=true
-datatypes/affix_fuzzer20.py linguist-generated=true
-datatypes/affix_fuzzer21.py linguist-generated=true
-datatypes/affix_fuzzer3.py linguist-generated=true
-datatypes/affix_fuzzer4.py linguist-generated=true
-datatypes/affix_fuzzer5.py linguist-generated=true
-datatypes/flattened_scalar.py linguist-generated=true
-datatypes/primitive_component.py linguist-generated=true
-datatypes/string_component.py linguist-generated=true
