+++ rerun_py/tests/test_types/archetypes/.gitattributes
@@ -1,0 +1,8 @@
+# DO NOT EDIT! This file is generated by crates/re_types_builder/src/lib.rs
+
+.gitattributes linguist-generated=true
+__init__.py linguist-generated=true
+affix_fuzzer1.py linguist-generated=true
+affix_fuzzer2.py linguist-generated=true
+affix_fuzzer3.py linguist-generated=true
+affix_fuzzer4.py linguist-generated=true
