+++ rerun_py/tests/test_types/blueprint/.gitattributes
@@ -1,0 +1,4 @@
+# DO NOT EDIT! This file is generated by crates/re_types_builder/src/lib.rs
+
+.gitattributes linguist-generated=true
+__init__.py linguist-generated=true
