+++ rerun_py/tests/test_types/components/.gitattributes
@@ -1,0 +1,25 @@
+# DO NOT EDIT! This file is generated by crates/re_types_builder/src/lib.rs
+
+.gitattributes linguist-generated=true
+__init__.py linguist-generated=true
+affix_fuzzer1.py linguist-generated=true
+affix_fuzzer10.py linguist-generated=true
+affix_fuzzer11.py linguist-generated=true
+affix_fuzzer12.py linguist-generated=true
+affix_fuzzer13.py linguist-generated=true
+affix_fuzzer14.py linguist-generated=true
+affix_fuzzer15.py linguist-generated=true
+affix_fuzzer16.py linguist-generated=true
+affix_fuzzer17.py linguist-generated=true
+affix_fuzzer18.py linguist-generated=true
+affix_fuzzer19.py linguist-generated=true
+affix_fuzzer2.py linguist-generated=true
+affix_fuzzer20.py linguist-generated=true
+affix_fuzzer21.py linguist-generated=true
+affix_fuzzer3.py linguist-generated=true
+affix_fuzzer4.py linguist-generated=true
+affix_fuzzer5.py linguist-generated=true
+affix_fuzzer6.py linguist-generated=true
+affix_fuzzer7.py linguist-generated=true
+affix_fuzzer8.py linguist-generated=true
+affix_fuzzer9.py linguist-generated=true
